% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuArc.R
\name{visuArc}
\alias{visuArc}
\title{VisuArc: An arcdiagram for VisuNet diagram}
\usage{
visuArc(net, decision, feature = NULL)
}
\arguments{
\item{net}{The saved VisuNet object}

\item{decision}{The decision for which you want to display the diagram.}

\item{feature}{The feature that you are interested to explore. By default is the hub node.}
}
\value{
An arc diagram
}
\description{
VisuArc: An arcdiagram for VisuNet diagram
}
\examples{
# ros<-rosetta(autcon)
# vis<-visunet(ros$main)
# visuArc(net= vis,decision= 'autism',feature='ZSCAN18')
}
